replaced all calls to check alpha blending mode of OpenGL JOGL graphics with built-in method of EventFilter now.
git-svn-id: https://svn.code.sf.net/p/jaer/code/jAER/trunk@6026 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/ch/unizh/ini/jaer/projects/davis/frames/ApsFrameExtractor.java
@@ -378,7 +378,7 @@ public class ApsFrameExtractor extends EventFilter2D implements Observer /* Obse
         }
     }
 
-    public void setDisplayFrameRGB(double[] frame) {
+    public void setDisplayFrameRGB(float[] frame) {
         int xc = 0;
         int yc = 0;
         for (int i = 0; i < frame.length; i += 3) {
